# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Utiva Growth AI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  enrolled: "enrolled",
  lost: "lost",
} as const;

export type LeadSource = (typeof LeadSource)[keyof typeof LeadSource];

export const LeadSource = {
  website: "website",
  referral: "referral",
  social_media: "social_media",
  event: "event",
  paid_ad: "paid_ad",
  organic: "organic",
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone?: string;
  courseInterest: string;
  status: LeadStatus;
  source: LeadSource;
  notes?: string;
  /** AI-generated lead quality score (0-100) */
  aiScore: number;
  /** AI recommendation for next action */
  aiSuggestedAction?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadBodySource =
  (typeof CreateLeadBodySource)[keyof typeof CreateLeadBodySource];

export const CreateLeadBodySource = {
  website: "website",
  referral: "referral",
  social_media: "social_media",
  event: "event",
  paid_ad: "paid_ad",
  organic: "organic",
} as const;

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string;
  courseInterest: string;
  source: CreateLeadBodySource;
  notes?: string;
}

export type UpdateLeadBodyStatus =
  (typeof UpdateLeadBodyStatus)[keyof typeof UpdateLeadBodyStatus];

export const UpdateLeadBodyStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  enrolled: "enrolled",
  lost: "lost",
} as const;

export interface UpdateLeadBody {
  status?: UpdateLeadBodyStatus;
  notes?: string;
  courseInterest?: string;
}

export type EnrollmentStatus =
  (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus];

export const EnrollmentStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  dropped: "dropped",
} as const;

export interface Enrollment {
  id: number;
  leadId?: number;
  studentName: string;
  email: string;
  course: string;
  cohort: string;
  status: EnrollmentStatus;
  amountPaid: number;
  enrolledAt: string;
  completionRate: number;
  /** AI-predicted engagement score */
  aiEngagementScore: number;
}

export interface CreateEnrollmentBody {
  leadId?: number;
  studentName: string;
  email: string;
  course: string;
  cohort: string;
  amountPaid: number;
}

export type UpdateEnrollmentBodyStatus =
  (typeof UpdateEnrollmentBodyStatus)[keyof typeof UpdateEnrollmentBodyStatus];

export const UpdateEnrollmentBodyStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  dropped: "dropped",
} as const;

export interface UpdateEnrollmentBody {
  status?: UpdateEnrollmentBodyStatus;
  completionRate?: number;
}

export type WorkflowType = (typeof WorkflowType)[keyof typeof WorkflowType];

export const WorkflowType = {
  lead_followup: "lead_followup",
  enrollment_reminder: "enrollment_reminder",
  re_engagement: "re_engagement",
  content_push: "content_push",
  lead_scoring: "lead_scoring",
  drop_alert: "drop_alert",
} as const;

export interface Workflow {
  id: number;
  name: string;
  description: string;
  type: WorkflowType;
  isActive: boolean;
  triggerCondition: string;
  actionTemplate: string;
  runsTotal: number;
  runsThisWeek: number;
  successRate: number;
  lastRunAt?: string;
  createdAt: string;
}

export type CreateWorkflowBodyType =
  (typeof CreateWorkflowBodyType)[keyof typeof CreateWorkflowBodyType];

export const CreateWorkflowBodyType = {
  lead_followup: "lead_followup",
  enrollment_reminder: "enrollment_reminder",
  re_engagement: "re_engagement",
  content_push: "content_push",
  lead_scoring: "lead_scoring",
  drop_alert: "drop_alert",
} as const;

export interface CreateWorkflowBody {
  name: string;
  description: string;
  type: CreateWorkflowBodyType;
  triggerCondition: string;
  actionTemplate: string;
}

export interface WorkflowRunResult {
  workflowId: number;
  affectedCount: number;
  status: string;
  message: string;
  /** Sample AI-generated message produced by the workflow */
  aiGeneratedMessage?: string;
}

export interface AnalyticsSummary {
  totalLeads: number;
  newLeadsThisWeek: number;
  totalEnrollments: number;
  enrollmentsThisWeek: number;
  conversionRate: number;
  totalRevenue: number;
  revenueThisMonth: number;
  activeWorkflows: number;
  workflowRunsThisWeek: number;
  avgAiLeadScore: number;
  topPerformingCourse: string;
}

export interface FunnelStage {
  stage: string;
  count: number;
  percentage: number;
}

export interface TrendPoint {
  week: string;
  enrollments: number;
  leads: number;
  revenue: number;
}

export interface Course {
  id: number;
  name: string;
  category: string;
  duration: string;
  price: number;
  enrollmentCount: number;
  completionRate: number;
  rating: number;
}

export type ContentItemType =
  (typeof ContentItemType)[keyof typeof ContentItemType];

export const ContentItemType = {
  email: "email",
  social_post: "social_post",
  whatsapp: "whatsapp",
  blog: "blog",
  ad_copy: "ad_copy",
} as const;

export type ContentItemStatus =
  (typeof ContentItemStatus)[keyof typeof ContentItemStatus];

export const ContentItemStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
} as const;

export interface ContentItem {
  id: number;
  title: string;
  body: string;
  type: ContentItemType;
  targetCourse?: string;
  targetAudience?: string;
  status: ContentItemStatus;
  scheduledFor?: string;
  aiGenerated: boolean;
  createdAt: string;
}

export type GenerateContentBodyType =
  (typeof GenerateContentBodyType)[keyof typeof GenerateContentBodyType];

export const GenerateContentBodyType = {
  email: "email",
  social_post: "social_post",
  whatsapp: "whatsapp",
  blog: "blog",
  ad_copy: "ad_copy",
} as const;

export type GenerateContentBodyTone =
  (typeof GenerateContentBodyTone)[keyof typeof GenerateContentBodyTone];

export const GenerateContentBodyTone = {
  professional: "professional",
  casual: "casual",
  urgent: "urgent",
  inspirational: "inspirational",
} as const;

export interface GenerateContentBody {
  type: GenerateContentBodyType;
  targetCourse: string;
  targetAudience?: string;
  tone?: GenerateContentBodyTone;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  workflow_run: "workflow_run",
  lead_created: "lead_created",
  enrollment_created: "enrollment_created",
  content_generated: "content_generated",
  lead_scored: "lead_scored",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  workflowName?: string;
  affectedEntity?: string;
  timestamp: string;
}

export type ListLeadsParams = {
  status?: ListLeadsStatus;
  course?: string;
  limit?: number;
};

export type ListLeadsStatus =
  (typeof ListLeadsStatus)[keyof typeof ListLeadsStatus];

export const ListLeadsStatus = {
  new: "new",
  contacted: "contacted",
  qualified: "qualified",
  enrolled: "enrolled",
  lost: "lost",
} as const;

export type ListEnrollmentsParams = {
  status?: ListEnrollmentsStatus;
  course?: string;
};

export type ListEnrollmentsStatus =
  (typeof ListEnrollmentsStatus)[keyof typeof ListEnrollmentsStatus];

export const ListEnrollmentsStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  dropped: "dropped",
} as const;

export type ListContentParams = {
  status?: ListContentStatus;
};

export type ListContentStatus =
  (typeof ListContentStatus)[keyof typeof ListContentStatus];

export const ListContentStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
} as const;

export type ListActivityParams = {
  limit?: number;
};