# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Utiva Growth AI Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, enrolled, lost]
        - name: course
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a single lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List all enrollments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, completed, dropped]
        - name: course
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Create a new enrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentBody"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    patch:
      operationId: updateEnrollment
      tags: [enrollments]
      summary: Update enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnrollmentBody"
      responses:
        "200":
          description: Updated enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List AI-generated content calendar items
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, published]
      responses:
        "200":
          description: Content items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
  /content/generate:
    post:
      operationId: generateContent
      tags: [content]
      summary: Use AI to generate a new content idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentBody"
      responses:
        "201":
          description: Generated content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        courseInterest:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, enrolled, lost]
        source:
          type: string
          enum: [website, referral, social_media, event, paid_ad, organic]
        notes:
          type: string
        aiScore:
          type: integer
          description: AI-generated lead quality score (0-100)
        aiSuggestedAction:
          type: string
          description: AI recommendation for next action
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - courseInterest
        - status
        - source
        - aiScore
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        courseInterest:
          type: string
        source:
          type: string
          enum: [website, referral, social_media, event, paid_ad, organic]
        notes:
          type: string
      required:
        - name
        - email
        - courseInterest
        - source
    UpdateLeadBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, qualified, enrolled, lost]
        notes:
          type: string
        courseInterest:
          type: string
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        studentName:
          type: string
        email:
          type: string
        course:
          type: string
        cohort:
          type: string
        status:
          type: string
          enum: [pending, active, completed, dropped]
        amountPaid:
          type: number
        enrolledAt:
          type: string
          format: date-time
        completionRate:
          type: integer
        aiEngagementScore:
          type: integer
          description: AI-predicted engagement score
      required:
        - id
        - studentName
        - email
        - course
        - cohort
        - status
        - amountPaid
        - enrolledAt
        - completionRate
        - aiEngagementScore
    CreateEnrollmentBody:
      type: object
      properties:
        leadId:
          type: integer
        studentName:
          type: string
        email:
          type: string
        course:
          type: string
        cohort:
          type: string
        amountPaid:
          type: number
      required:
        - studentName
        - email
        - course
        - cohort
        - amountPaid
    UpdateEnrollmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, completed, dropped]
        completionRate:
          type: integer
...
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [email, social_post, whatsapp, blog, ad_copy]
        targetCourse:
          type: string
        targetAudience:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
        scheduledFor:
          type: string
          format: date-time
        aiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - body
        - type
        - status
        - aiGenerated
        - createdAt
    GenerateContentBody:
      type: object
      properties:
        type:
          type: string
          enum: [email, social_post, whatsapp, blog, ad_copy]
        targetCourse:
          type: string
        targetAudience:
          type: string
        tone:
          type: string
          enum: [professional, casual, urgent, inspirational]
      required:
        - type
        - targetCourse
...