# lib/api-zod/src/generated/types/createEnrollmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Utiva Growth AI Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEnrollmentBody {
  leadId?: number;
  studentName: string;
  email: string;
  course: string;
  cohort: string;
  amountPaid: number;
}