# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Utiva Growth AI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeadBodySource } from "./createLeadBodySource";

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string;
  courseInterest: string;
  source: CreateLeadBodySource;
  notes?: string;
}