# lib/api-zod/src/generated/types/updateLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Utiva Growth AI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeadBodyStatus } from "./updateLeadBodyStatus";

export interface UpdateLeadBody {
  status?: UpdateLeadBodyStatus;
  notes?: string;
  courseInterest?: string;
}